refactor: move authn providers to adapters
